webdev/backend/requirements.txt
awsiot==0.1.3
cachetools==5.3.3
fastapi==0.135.1
msgpack_python==0.5.6
protobuf==7.34.0
python-dotenv==1.2.2
